feat: restructure wip
