+++ packages/api/.env.sample
@@ -1,2 +1,2 @@
 PORT=<EXPRESS_PORT>
-DATABASE_URL=<DATABASE_URL>
+DATABASE_URL=postgresql://root:root@localhost:5432/grants-round?schema=public
