+++ packages/api/package.json
@@ -12,12 +12,12 @@
     "generate": "npx prisma generate --schema=./src/prisma/schema.prisma"
   },
   "devDependencies": {
+    "@prisma/client": "^4.7.0",
     "@pulumi/aws-quickstart-aurora-postgres": "^0.0.3",
     "@types/aws-lambda": "^8.10.108",
     "@types/express": "^4.17.14",
     "@types/node": "^14.18.33",
     "@types/node-fetch": "^2.6.2",
-    "@prisma/client": "^4.7.0",
     "concurrently": "^7.6.0",
     "jest": "^29.3.1",
     "nodemon": "^2.0.20",
@@ -31,7 +31,6 @@
     "@pulumi/awsx": "^1.0.0-beta",
     "@pulumi/pulumi": "^3.0.0",
     "aws-lambda": "^1.0.7",
-    "body-parser": "^1.20.1",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.17.1",
