+++ packages/api/src/prisma/schema.prisma
@@ -5,7 +5,7 @@ generator client {
 
 generator docs {
   provider = "node node_modules/prisma-docs-generator"
-  output = "../../docs"
+  output = "../../docs/prisma"
 }
 
 datasource db {
